requirements.txt
# Core Data Science
pandas>=1.5.0
numpy>=1.21.0
scikit-learn>=1.0.0
scipy>=1.7.0

# Data Visualization
matplotlib>=3.5.0
seaborn>=0.11.0
plotly>=5.0.0

# Jupyter & Development
jupyter>=1.0.0
ipykernel>=6.0.0
notebook>=6.4.0

# Model Serialization
joblib>=1.1.0

# Configuration
PyYAML>=6.0

# Testing
pytest>=7.0.0

# Optional: Advanced ML
xgboost>=1.5.0
lightgbm>=3.3.0